Removed warning ros_image_texture.cpp (#1775)
(cherry picked from commit 443a171bf49741707379eb6f1d66ba443dde7a23)

+++ rviz_default_plugins/src/rviz_default_plugins/displays/image/ros_image_texture.cpp
@@ -226,7 +226,7 @@ bool ROSImageTexture::update()
       data_size = repacked.size();
       // Reflect the repack so any downstream code reading stride_ sees the
       // new (packed) layout.
-      stride_ = packed_row;
+      stride_ = static_cast<uint32_t>(packed_row);
     }
   }
